refactor(system): package up system into openjarvis/system/
Move the four system-composition modules into a package:
  openjarvis/system.py          -> openjarvis/system/core.py
  openjarvis/system_bundles.py  -> openjarvis/system/bundles.py
  openjarvis/system_protocols.py -> openjarvis/system/protocols.py
  openjarvis/query_orchestrator.py -> openjarvis/system/orchestrator.py

__init__.py re-exports JarvisSystem, SystemBuilder, QueryOrchestrator,
the bundles, and OrchestratorDeps, so existing 'from openjarvis.system
import JarvisSystem' callers keep working. Update one test's logger
patch target to openjarvis.system.core.logger.

Also strip historical/issue-reference comments from the new modules.
