+++ src/openjarvis/system/__init__.py
@@ -1,0 +1,22 @@
+"""Top-level system composition: JarvisSystem, SystemBuilder, and helpers."""
+
+from openjarvis.system.bundles import (
+    AgentRuntime,
+    Observability,
+    Scheduling,
+    SecurityContext,
+)
+from openjarvis.system.core import JarvisSystem, SystemBuilder
+from openjarvis.system.orchestrator import QueryOrchestrator
+from openjarvis.system.protocols import OrchestratorDeps
+
+__all__ = [
+    "AgentRuntime",
+    "JarvisSystem",
+    "Observability",
+    "OrchestratorDeps",
+    "QueryOrchestrator",
+    "Scheduling",
+    "SecurityContext",
+    "SystemBuilder",
+]
